Rock paper scissors game
Setup the basic logic of the game.
Counts score of both players and
iterates the game a set amount of times.
My first time experimenting with function
calling.
Plan to make the game less hard-coded in
the future.
Learning more about objects and their uses,
have an idea to shorten the if/else comparisons.